+++ src/sps/completion.py
@@ -3,6 +3,25 @@ import os
 
 
 def get(cachefile, shell=None):
+    """Generates a shell completion script
+
+    Parameters
+    ----------
+    cachefile: str
+        Path to cache file
+    shell: str, optional
+        Name of the shell to generate completion script for
+        if not provided the function will try to use the 
+        SHELL environment varable
+
+    Raises
+    ------
+    SystemExit
+        if an unknown shell is provided
+        if shell is not provided and SHELL environ is not set
+        if completion generation file is not readable
+    """
+
     if shell == None:
         try:
             os.environ["SHELL"]